+++ packages/files/src/images/index.ts
@@ -1,0 +1,9 @@
+export {
+  renderImage,
+  UnsupportedImageError,
+  type Derivative,
+  type DerivativeSpec,
+  type ImageFormat,
+  type RenderedImage,
+  type RenderOptions,
+} from './render';